refactor legacy.cpp: extract DrsBkg, DRS and palette
